# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the comeht server api.",
        "title": "comeht Server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "A route for user login",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login for a user",
                "parameters": [
                    {
                        "description": "userRequest",
                        "name": "userRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.LoginResponse"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "A route to logout a user by cookie",
                "tags": [
                    "User"
                ],
                "summary": "Logout a user",
                "responses": {}
            }
        },
...
        "/register": {
            "post": {
                "description": "A route to register a new user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "userRequest",
                        "name": "userRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.LoginResponse"
                        }
                    }
                }
            }
        },
        "/results": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Route to get all results",
                "tags": [
                    "Results"
                ],
                "summary": "Get all results",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "A route to save a duel results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Duel"
                ],
                "summary": "Save the duel results",
                "parameters": [
                    {
                        "description": "results",
                        "name": "results",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.ResultsRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "views.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "views.Options": {
            "type": "object",
            "properties": {
                "aspect": {
                    "type": "string"
                },
                "choice_time": {
                    "type": "number"
                },
                "chosen_option": {
                    "type": "string"
                }
            }
        },
        "views.ResultsRequest": {
            "type": "object",
            "properties": {
                "duel_id": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Options"
                    }
                }
            }
        },
        "views.UserRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}